Clear warning backlog before latest release
The main release branch still emitted clippy warnings after the Merica and shortcut fixes were applied. This keeps the release tip warning-clean with mechanical clippy simplifications plus the few manual lints that clippy could not safely rewrite.

Constraint: User requested the latest release after warnings were fixed.

Rejected: Publish directly from the dirty existing main worktree | it contains unrelated local edits and would make the release ambiguous.

Rejected: Merge develop wholesale into main | develop is older release topology and would delete the current npm release automation.

Confidence: high

Scope-risk: moderate

Directive: Keep cargo clippy --all-targets warning-clean before future release tags.

Tested: cargo clippy --all-targets

Tested: cargo test --all-targets

Tested: cargo fmt --check

Tested: git diff --check

+++ tests/navigation_test.rs
@@ -344,7 +344,7 @@ async fn test_series_episodes_page_navigation() {
         app.series_episodes
             .push(matrix_iptv_lib::api::SeriesEpisode {
                 id: Some(FlexId::from_number(i as i64)),
-                episode_num: i as i32,
+                episode_num: i,
                 title: Some(format!("Episode {}", i)),
                 container_extension: Some("mp4".to_string()),
                 info: None,
